Replace Circulation Controller with Resource
Move Book, BookController, and CirculationController to Resource and
ResourceController. Re-factor tests and module structure.
