+++ web/controllers/check_out_controller.ex
@@ -5,7 +5,7 @@ defmodule Bookish.CheckOutController do
   plug Bookish.Plugs.RequireAuth when action in [:new, :create]
 
   alias Bookish.CheckOut
-  alias Bookish.Circulation
+  alias Bookish.Book
   alias Bookish.AuthController, as: Auth
 
   def index(conn, _params) do
@@ -18,7 +18,7 @@ defmodule Bookish.CheckOutController do
     check_out_params = Map.merge(data, %{"borrower_name" => user.name, "borrower_id" => user.id})
     changeset = 
       conn
-      |> Circulation.get_book_with_location
+      |> Bookish.BookController.update_with_location
       |> build_assoc(:check_outs)
       |> CheckOut.changeset(check_out_params)
 
@@ -37,7 +37,7 @@ defmodule Bookish.CheckOutController do
   defp assign_book(conn, _opts) do
     case conn.params do
       %{"book_id" => book_id} ->
-        book = Repo.get(Bookish.Book, book_id)
+        book = Repo.get(Book, book_id)
         assign(conn, :book, book)
       {:ok, _book} ->
         conn
